producer/pyproject.toml
[project]
name = "sales-order-producer"
version = "0.1.0"
description = "Kafka producer for composite sales orders"
requires-python = ">=3.11"
dependencies = [
    "kafka-python==2.0.2",
]

[project.scripts]
producer = "app.main:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
